Rename ia_model.dart and update import paths
Renamed lib/class/ia_model.dart to lib/classes/ia_model.dart and updated the import statement in gemini_service.dart to reflect the new path.
