+++ lib/classes/ia_model.dart
